+++ dockers/squid/run.sh
@@ -1,11 +1,13 @@
 #!/bin/bash
 
 if [ ! "$SPICE_HOSTS" == "false" ]; then
-	$HOSTS=$(echo $SPICE_HOSTS |tr "," " ")
+	HOSTS=$(echo $SPICE_HOSTS |tr "," " ")
         sed -i "/^acl SPICE_HOSTS/c\acl SPICE_HOSTS dst $HOSTS" /etc/squid/squid.conf
-        sed -i "/^http port/c\http port $SPICE_PROXY_PORT" /etc/squid/squid.conf
+        sed -i "/^http_port/c\http_port $SPICE_PROXY_PORT" /etc/squid/squid.conf
         squid -N
 else
-	echo "No proxy config found"
+	echo "SPICE_HOSTS=false in .env"
+	echo "So exitting squid instance as not needed."
+	echo "Access to viewer ports will be direct."
 	exit 1
 fi