+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 - Hypervisor autodetects VLANS or can be set by script.
 - Adding remote hypervisor through CLI has all the UI options now.
   docker exec -e ENABLED=True -e ID=isard-hypervisor -e HYPERVISOR=192.168.0.10 -e PASSWORD=isard -e PORT=2022 -e POOL=default -e DISKOP=True -e VIRTUALOP=True -e VIEWERHOST=localhost -e VIEWERNATHOST=isardvdi.com -e VIEWERNATOFFSET=30000 isard-app sh -c '/add-hypervisor.sh'
+- Added squid image that allows proxying spice viewers.
 - Now .env file allows to configure main system parameters.
 - Grafana container updated to latest version and uses .env file
 - Rsync added to Hypervisor image so now can handle different mount points.